[refactor] 알림 상태 ENUM class 위치 변경
