+++ Member/src/main/java/com/rainbowgon/member/global/entity/NotificationStatus.java
@@ -1,4 +1,4 @@
-package com.rainbowgon.member.domain.profile.entity;
+package com.rainbowgon.member.global.entity;
 
 public enum NotificationStatus {
     ON,
